start refactoring command matches into trait
