+++ mini-projects/00-jq-rs/src/maps/array_slice.rs
@@ -2,19 +2,18 @@ use serde_json::Value;
 use crate::maps::maps::Map;
 use anyhow::Result;
 
-pub struct ObjectIdentifierMap {
-    pub key: String,
+pub struct ArrayIteratorMap {
+    pub from: usize,
+    pub to: usize,
 }
 
-impl Map for ObjectIdentifierMap {
+impl Map for ArrayIteratorMap {
     fn map(&self, value: Result<Vec<Value>>) -> Result<Vec<Value>> {
         let value = value?;
         let result: Result<Vec<Value>> = value.iter().map(|v| {
-            let new_value = v[&self.key].clone();
-            if new_value.is_null() {
-                anyhow::bail!("object identifier not found")
-            }
-            Ok(new_value)
+            let array = v.as_array().ok_or_else(|| anyhow::anyhow!("array iterator requires an array"))?;
+            let slice = array[self.from..self.to].to_vec();
+            Ok(Value::Array(slice))
         }).collect();
         result
     }
@@ -41,3 +40,4 @@ impl Map for ObjectIdentifierMap {
         }
     }
 }
+
