+++ mini-projects/00-jq-rs/src/maps/del.rs
@@ -10,5 +10,20 @@ impl Map for DelMap {
     fn map(&self, value: Result<Vec<Value>>) -> Result<Vec<Value>> {
         value
     }
+
+    fn command_match(&self, input: &str) -> Result<Box<dyn Map>> {
+        let pattern = r"del\(\.(\w+)\)";
+        let re: &'static Regex = &Regex::new(pattern).unwrap();
+    
+        match re.captures(input) {
+            Some(captures) => {
+                let key = captures.get(1).unwrap().as_str();
+                Ok(Box::new(DelMap {
+                    key: key.to_string(),
+                }))
+            }
+            None => anyhow::bail!("failed to parse del slice"),
+        }
+    }
 }
