+++ mini-projects/00-jq-rs/src/maps/maps.rs
@@ -3,6 +3,8 @@ use anyhow::Result;
 
 pub trait Map {
     fn map(&self, value: Result<Vec<Value>>) -> Result<Vec<Value>>;
+
+    fn command_match(&self, input: &str) -> Result<Box<dyn Map>>;
 }
